+++ src/main/java/com/skyflow/errors/ErrorCode.java
@@ -36,7 +36,16 @@ public enum ErrorCode {
     InvalidConnectionInput(400, "Invalid connection Input"),
     EmptyJSONString(400, "credentials string cannot be empty"),
     InvalidJSONStringFormat(400, "credentials string is not a valid json string format"),
-    EmptyFilePath(400, "file path cannot be empty or null");
+    EmptyFilePath(400, "file path cannot be empty or null"),
+    EmptyContext(400, "ctx claim field is missing from the jwt assertion"),
+    IncorrectRole(400,"Requested scope  cannot be granted"),
+
+    IncorrectCredentials(400,"Incorrect credentials provided"),
+    InvalidUpsertOptionType(400,"upsert options should be an non empty UpsertOption array."),
+    InvalidUpsertObjectType(400,"upsert option cannot be null, should be an UpsertOption object."),
+    InvalidTableInUpsertOption(400,"Invalid table in upsert object, non empty string is required."),
+    InvalidColumnInUpsertOption(400,"Invalid column in upsert object, non empty string is required.");
+
     private final int code;
     private final String description;